🔧 Add boolalpha

+++ lab1/task2.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 int main() {
-  int a = 5, b = 3;
+  short a = 5, b = 3;
 
   std::cout << "Liczba a: " << a << std::endl;
   std::cout << "Liczba b: " << b << std::endl;
